% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene_frequency.R
\name{gene_frequency}
\alias{gene_frequency}
\title{Run sparse PCA Using Quasi-poisson pearson residuals (skip dense normalization)}
\usage{
gene_frequency(
  x,
  obs = NULL,
  batch_variable = NULL,
  cellid_colname = "cell_ID"
)
}
\arguments{
\item{x}{A sparse genes x cells counts matrix}

\item{obs}{A data frame of metadata information, containing columns for the 'batch variable' and 'cell id'.
Or, row names of the data frame indicating 'cell id'.}

\item{batch_variable}{column in \code{obs} indicating 'batch'.  i.e., 'patient' or 'samplename' or 'slide_id'.}

\item{cellid_colname}{column in \code{obs} indicating the cell identifier, i.e., 'cell_ID'.}
}
\description{
Recreate Seurat RunPCA output from pearson residual normalization (described in Lause 2021),
WITHOUT creating a dense cells x genes scaled data matrix.
This function uses Quasi-poisson instead of Negative Binomial distribution for computational convenience.
Computes PCA via SVD decomposition of genes x genes matrix instead of
SVD decomposition of cells x genes dense matrix, which is useful for memory when cells x genes >> genes^2 .
}
