% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_umap.R
\name{make_umap}
\alias{make_umap}
\title{A UMAP helper function}
\usage{
make_umap(
  pcaobj,
  min_dist = 0.01,
  n_neighbors = 30,
  metric = "cosine",
  key = "UMAP_"
)
}
\arguments{
\item{pcaobj}{A pca list object containing a DimReduc object called 'reduction.data' which is passed to \code{uwot::umap}. Same format as that returned by \code{sparse_quasipoisson_pca_seurat}.}

\item{min_dist}{hyperparameter passed directly to \code{uwot::umap}}

\item{n_neighbors}{hyperparameter passed directly to \code{uwot::umap}}

\item{metric}{hyperparameter passed directly to \code{uwot::umap}}

\item{key}{DimReduc key for umap object.
subset (of, say, highly variable) genes to be used in PCA.}
}
\description{
Create a UMAP from a PCA object and also return the fuzzy nearest-neighbors graph used to
build the UMAP.  This nearest neighbor graph can be used as the starting point for unsupervised clustering,
potentially improving the agreement between UMAP and unsupervised clusters derived from the same PC data.
}
