% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlation_of_pearson.R
\name{pearson_cormat}
\alias{pearson_cormat}
\title{Compute the gene x gene correlation matrix of quasi-poisson Pearson residuals}
\usage{
pearson_cormat(
  x,
  totalcounts = NULL,
  grate = NULL,
  quasi_poisson_variance_inflation = 1.01
)
}
\arguments{
\item{x}{A sparse genes x cells counts matrix}

\item{totalcounts}{The total UMI counts across all gene targets in each cell.  This can be pre-computed by the user and passed in cases where the counts matrix \code{x} contains only a
subset (of, say, highly variable) genes to be used in PCA.}

\item{grate}{The gene frequencies.  This can be pre-computed by the user and passed in cases where the counts matrix \code{x} contains only a
subset (of, say, highly variable) genes to be used in PCA.}

\item{quasi_poisson_variance_inflation}{factor of overdispersion for Variance relative to the mean.  i.e., V(x) = E(x) * phi, where phi is the variance inflation factor}
}
\value{
A dense genes x genes correlation matrix.
}
\description{
Computes the (dense) genes x genes Pearson correlation matrix of the quasi-poisson Pearson
residuals for counts matrix \code{x}, reusing the same genes x genes crossproduct algebra as
\code{sparse_quasipoisson_pca_seurat()}, WITHOUT ever forming the (dense) cells x genes residuals matrix.
}
