% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pearson_normalize_gamma_glm.R
\name{pearson_normalize_gamma_glm}
\alias{pearson_normalize_gamma_glm}
\title{Normalize continuous protein expression data using pearson residuals}
\usage{
pearson_normalize_gamma_glm(x, lower_q_thresh = 0.01, upper_q_thresh = NULL)
}
\arguments{
\item{x}{A proteins x cells expression matrix}

\item{lower_q_thresh}{lower quantile at which to threshold raw expression for a given protein (default = 0.01) before fitting gamma glm model.
Note that for a gamma family glm used here, the protein expression must be positive.}

\item{upper_q_thresh}{upper quantile at which to threshold expression for a given protein (none by default) before fitting gamma glm model.
To avoid outliers skewing normalization, could threshold extreme values of raw protein expression}
}
\description{
Fits a gamma generalized linear model for each protein / row in the expression matrix \code{x},
and returns the pearson residuals.  This new proteins x cells matrix can be used as a normalized dataset for
downstream analyses.
}
