% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pearson_signatures.R
\name{pearson_signatures}
\alias{pearson_signatures}
\title{Compute gene signature scores as linear combinations of quasi-poisson Pearson residuals}
\usage{
pearson_signatures(
  x,
  wts,
  totalcounts = NULL,
  grate = NULL,
  quasi_poisson_variance_inflation = 1.01
)
}
\arguments{
\item{x}{A sparse genes x cells counts matrix}

\item{wts}{A genes x signatures matrix of weights, with rownames giving the gene names used in each signature}

\item{totalcounts}{The total UMI counts across all gene targets in each cell.  This can be pre-computed by the user and passed in cases where the counts matrix \code{x} contains only a
subset (of, say, highly variable) genes to be used in PCA.}

\item{grate}{The gene frequencies.  This can be pre-computed by the user and passed in cases where the counts matrix \code{x} contains only a
subset (of, say, highly variable) genes to be used in PCA.}

\item{quasi_poisson_variance_inflation}{factor of overdispersion for Variance relative to the mean.  i.e., V(x) = E(x) * phi, where phi is the variance inflation factor}
}
\value{
A cells x signatures matrix of signature scores.
}
\description{
For a set of gene signatures (weighted linear combinations of genes, given by \code{wts}), computes the
cell-level score \code{sum_g w_g * r_gc}, where \code{r_gc} is the quasi-poisson Pearson residual of gene g in
cell c, WITHOUT ever forming the (dense) Pearson residuals matrix.
}
