% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_umap.R
\name{plot_umap}
\alias{plot_umap}
\title{A UMAP plotting utility function}
\usage{
plot_umap(
  umapreduc,
  clustercol,
  semuse,
  cls = NULL,
  plotfirst = NULL,
  alpha = 1,
  max.overlaps = 100,
  lblsize = 4,
  cellid_colname = "cell_ID"
)
}
\arguments{
\item{umapreduc}{name of umap dimension reduction in a \code{semuse} seurat object.}

\item{clustercol}{name of column to color by in \code{semuse@meta.data}.}

\item{cls}{optional (possibly named) vector of colors to plot categories of \code{clustercol} by.}

\item{plotfirst}{optional vector of categories in \code{clustercol} to be plotted first
(potentially allowing other less frequent categories to be highlighted by overlaying on top).}

\item{alpha}{transparency of points to be plotted}

\item{max.overlaps}{max.overlaps parameter passed to \code{ggrepel::geom_label_repel}}

\item{lblsize}{size of labels to be passed to \code{ggrepel::geom_label_repel}}
}
\description{
A helper function for plotting UMAP using ggplot2 and ggrepel packages.
}
