% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project_sparse_quasipoisson_pca_seurat.R
\name{project_sparse_quasipoisson_pca_seurat}
\alias{project_sparse_quasipoisson_pca_seurat}
\title{Project PCA onto new data}
\usage{
project_sparse_quasipoisson_pca_seurat(
  x,
  pcaobj,
  ncores = 1,
  verbose = TRUE,
  block.size = 10^4,
  ndims.print = 1:5,
  nfeatures.print = 30,
  return_seurat_reduction = TRUE
)
}
\arguments{
\item{x}{A sparse genes x cells counts matrix}

\item{pcaobj}{generated from \code{sparse_quasipoisson_pca_seurat}
Analogous to Seurat::ScaleData scale.max argument, with same default of "10".}

\item{ncores}{the # of cores to use to project embeddings}

\item{verbose}{print out seurat-style PCA summary}

\item{block.size}{the # of cells to project into PCA space at a time. Can reduce block.size for lower-memory computing.}

\item{ndims.print}{the # of first PC's to print for verbose-seurat-style printout.}

\item{nfeatures.print}{the # of top genes to show per PC for verbose-seurat-style printout.}

\item{return_seurat_reduction}{if TRUE, returns a reduction.data object in Seurat format.}
}
\description{
Take the results generated from \code{sparse_quasipoisson_pca_seurat} and project them onto new cells.
}
