% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sparse_quasipoisson_pca_seurat_batch.R
\name{sparse_quasipoisson_pca_seurat_batch}
\alias{sparse_quasipoisson_pca_seurat_batch}
\title{Run sparse PCA Using Quasi-poisson pearson residuals computed by batch (skip dense normalization)}
\usage{
sparse_quasipoisson_pca_seurat_batch(
  x,
  obs,
  batch_variable = NULL,
  cellid_colname = "cell_ID",
  scale.max = 10,
  reduction.key = "PC_",
  do.scale = FALSE,
  do.center = FALSE,
  totalcounts = NULL,
  grate = NULL,
  quasi_poisson_variance_inflation = 1.01,
  npcs = 50,
  ncores = 1,
  verbose = TRUE,
  block.size = 10^4,
  ndims.print = 1:5,
  nfeatures.print = 30,
  return_seurat_reduction = TRUE
)
}
\arguments{
\item{x}{A sparse genes x cells counts matrix}

\item{obs}{a data.frame or data.table of metadata, containing the 'batch variable' and the 'cell id'.}

\item{batch_variable}{name of column in \code{obs} data frame  corresponding to 'batch'.}

\item{cellid_colname}{name of column in \code{obs} data frame corresponding to 'cell id'.}

\item{scale.max}{clip expression more than 'scale.max' standard deviations from the mean.
Analogous to Seurat::ScaleData scale.max argument, with same default of "10".}

\item{reduction.key}{name of key for Seurat reduction object (if return_seurat_reduction=TRUE)}

\item{do.scale}{gene x gene cov matrix reflects as if pearson residuals are scaled by their standard deviation}

\item{do.center}{gene x gene cov matrix reflects as if the pearson residuals were centered by their mean}

\item{totalcounts}{The total UMI counts across all gene targets in each cell.  This can be pre-computed by the user and passed in cases where the counts matrix \code{x} contains only a
subset (of, say, highly variable) genes to be used in PCA.}

\item{grate}{The gene frequencies.  This can be pre-computed by the user and passed in cases where the counts matrix \code{x} contains only a
subset (of, say, highly variable) genes to be used in PCA.}

\item{quasi_poisson_variance_inflation}{factor of overdispersion for Variance relative to the mean.  i.e., V(x) = E(x) * phi, where phi is the variance inflation factor}

\item{npcs}{the # of PCs}

\item{ncores}{the # of cores to use to project embeddings}

\item{verbose}{print out seurat-style PCA summary}

\item{block.size}{the # of cells to project into PCA space at a time.}

\item{ndims.print}{the # of first PC's to print for verbose-seurat-style printout.}

\item{nfeatures.print}{the # of top genes to show per PC for verbose-seurat-style printout.}

\item{return_seurat_reduction}{if TRUE, returns a reduction.data object in Seurat format.}
}
\description{
Recreate Seurat RunPCA output from pearson residual normalization (described in Lause 2021),
WITHOUT creating a dense cells x genes scaled data matrix.
This function uses Quasi-poisson instead of Negative Binomial distribution for computational convenience.
Computes PCA via SVD decomposition of genes x genes matrix instead of
SVD decomposition of cells x genes dense matrix, which is useful for memory when cells x genes >> genes^2 .
}
